cc_project2/requirements.txt
Flask==2.3.3
Pillow==10.0.0
gunicorn==20.1.0
google-cloud-storage
